checkForStraights 1
Requirement:

Declare a checkForStraights function which accepts an array of numbers. If the user gets a large straight, update the fifth radio button with a score of 40. If the user gets a small straight, update the fourth radio button with a score of 30. If the user gets no straight, update the last radio button to display 0.

Call your checkForStraights function when the rollDiceBtn is clicked to complete your dice game!